requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.6.0
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.4
executing==0.10.0
fastjsonschema==2.16.1
frozenlist==1.3.0
great-expectations==0.15.20
grpcio==1.47.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.14.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
lxml==4.9.1
makefun==1.14.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.2
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.11
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.1
requests==2.28.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scipy==1.9.0
Send2Trash==1.8.0
six==1.16.0
slack-sdk==3.18.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
stack-data==0.4.0
termcolor==1.1.0
terminado==0.15.0
tinycss2==1.1.1
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.2
yarg==0.1.9
yarl==1.7.2
zipp==3.8.1
turbine-py>=1.1.0